long to bigint

+++ src/main/java/HOTPGenerator.java
@@ -1,6 +1,7 @@
 import interfaces.IHOTPGenerator;
 
 import java.io.UnsupportedEncodingException;
+import java.math.BigInteger;
 import java.net.URI;
 import java.net.URISyntaxException;
 
@@ -44,6 +45,6 @@ public class HOTPGenerator extends OneTimePasswordGenerator implements IHOTPGene
      */
     @Override
     public String generate(long counter) throws IllegalArgumentException {
-        return super.generate(counter);
+        return super.generate(BigInteger.valueOf(counter));
     }
 }
