+++ src/main/java/interfaces/ITOTPGenerator.java
@@ -1,6 +1,5 @@
 package interfaces;
 
-import java.math.BigInteger;
 import java.time.Duration;
 import java.time.Instant;
 import java.util.Date;
@@ -9,7 +8,7 @@ public interface ITOTPGenerator {
     String generate();
     String generate(Instant instant);
     String generate(Date date);
-    String generate(BigInteger secondsPast1970);
+    String generate(long secondsPast1970);
     boolean verify(String code);
     Duration getPeriod();
 }
